+++ controllers/agents.js
@@ -376,6 +376,43 @@ router.get('/outdated', cache(), function(req, res) {
 })
 
 /**
+ * @api {get} /agents/name/:agent_name Get an agent by its name
+ * @apiName GetAgentsName
+ * @apiGroup Info
+ *
+ * @apiParam {String} agent_name Agent name.
+ *
+ * @apiDescription Returns the information of an agent called :agent_name.
+ *
+ * @apiExample {curl} Example usage:
+ *     curl -u foo:bar -k -X GET "https://127.0.0.1:55000/agents/name/myAgent?pretty"
+ *
+ */
+router.get('/name/:agent_name', cache(), function(req, res) {
+    logger.debug(req.connection.remoteAddress + " GET /agents/name/:agent_name");
+
+    req.apicacheGroup = "agents";
+
+    var data_request = {'function': '/agents/name/:agent_name', 'arguments': {}};
+    var filters = {'select':'select_param'};
+
+    if (!filter.check(req.params, {'agent_name':'names'}, req, res))  // Filter with error
+        return;
+
+    data_request['arguments']['agent_name'] = req.params.agent_name;
+
+    if(!filter.check(req.query, filters, req, res)) // Filter with error
+        return;
+
+    if ('select' in req.query)
+        data_request['arguments']['select'] =
+        filter.select_param_to_json(req.query.select);
+
+    execute.exec(python_bin, [wazuh_control], data_request, function (data) { res_h.send(req, res, data); });
+
+})
+
+/**
  * @api {get} /agents/:agent_id Get an agent
  * @apiName GetAgentsID
  * @apiGroup Info